step 1: analysis
