+++ examples/src/old-bass-nation.cpp
@@ -1,9 +1,7 @@
 #include "ExampleFramework.hpp"
 #include <avz/Layer.hpp>
 #include <avz/SpectrumDrawable.hpp>
-#include <avz/fft/AudioAnalyzer.hpp>
-#include <avz/fft/FrequencyAnalyzer.hpp>
-#include <avz/fft/Interpolator.hpp>
+#include <avz/analysis.hpp>
 #include <avz/fx/Polar.hpp>
 #include <avz/util.hpp>
