+++ examples/src/stereo-scope.cpp
@@ -2,6 +2,7 @@
 #include "ExampleFramework.hpp"
 #include <avz/ScopeDrawable.hpp>
 #include <avz/util.hpp>
+#include <avz/analysis.hpp>
 
 #include <algorithm>
 #include <cmath>
