+++ libavz/analysis/include/avz/analysis/BinPacker.hpp
