+++ libavz/analysis/src/FrequencyAnalyzer.cpp
@@ -1,4 +1,4 @@
-#include <avz/fft/FrequencyAnalyzer.hpp>
+#include "FrequencyAnalyzer.hpp"
 #include <cassert>
 #include <cmath>
